refactor: rename callback out module for import useness
