+++ go.work
@@ -35,5 +35,5 @@ use(
     ./network
     ./_script
     ./renew
-    ./callback-out
+    ./callback_out
 )
